chore: add build tags to control login with Aqua being included
Not ready to release the aqua login code so putting behind a build tag
