+++ Makefile
@@ -2,12 +2,21 @@ SED=$(shell command -v gsed || command -v sed)
 PLATFORMS = linux/amd64 linux/arm64 darwin/amd64 darwin/arm64 windows/amd64
 OUTPUTS = $(patsubst %,%/trivy-mcp,$(PLATFORMS))
 
+BUILDTAGS ?=
+BUILD_FLAGS = -tags "$(BUILDTAGS)"
+
 .PHONY: clean
 clean:
 	rm -rf trivy-mcp*
 
 .PHONY: build
 build: clean $(OUTPUTS)
+
+# Build with Aqua support
+.PHONY: build-aqua
+build-aqua:
+	$(MAKE) build BUILDTAGS=aqua
+
 %/trivy-mcp:
 	@[ $$NEW_VERSION ] || ( echo "env 'NEW_VERSION' is not set"; exit 1 );
 	@trivy_version=$$(cat go.mod | grep 'github.com/aquasecurity/trivy v' | awk '{print $$2}'); \
@@ -19,7 +28,7 @@ build: clean $(OUTPUTS)
 	echo "Building for $*..."; \
 	GOOS=$(word 1,$(subst /, ,$*)); \
 	GOARCH=$(word 2,$(subst /, ,$*)); \
-	CGO_ENABLED=0 GOOS=$$GOOS GOARCH=$$GOARCH go build -ldflags "-s -w -X github.com/aquasecurity/trivy-mcp/pkg/version.Version=v$${NEW_VERSION} -X github.com/aquasecurity/trivy-mcp/pkg/version.TrivyVersion=$${trivy_version}" -o trivy-mcp ./cmd/trivy-mcp/main.go; \
+	CGO_ENABLED=0 GOOS=$$GOOS GOARCH=$$GOARCH go build $(BUILD_FLAGS) -ldflags "-s -w -X github.com/aquasecurity/trivy-mcp/pkg/version.Version=v$${NEW_VERSION} -X github.com/aquasecurity/trivy-mcp/pkg/version.TrivyVersion=$${trivy_version}" -o trivy-mcp ./cmd/trivy-mcp/main.go; \
 	if [ $$GOOS = "windows" ]; then \
 		mv trivy-mcp trivy-mcp.exe; \
 		tar -czf trivy-mcp-$$GOOS-$$GOARCH.tar.gz plugin.yaml trivy-mcp.exe LICENSE > /dev/null; \
@@ -46,15 +55,20 @@ install-plugin: add-plugin-manifest
 	@echo "Installing plugin..."
 	@trivy_version=$$(cat go.mod | grep 'github.com/aquasecurity/trivy v' | awk '{ print $$2}') ;\
 	echo Current trivy version: $$trivy_version ;\
-	go build -ldflags "-s -w -X github.com/aquasecurity/trivy-mcp/pkg/version.TrivyVersion=$${trivy_version}" -o ~/.trivy/plugins/mcp/trivy-mcp ./cmd/trivy-mcp/main.go;
+	go build $(BUILD_FLAGS) -ldflags "-s -w -X github.com/aquasecurity/trivy-mcp/pkg/version.TrivyVersion=$${trivy_version}" -o ~/.trivy/plugins/mcp/trivy-mcp ./cmd/trivy-mcp/main.go;
 	@echo "Plugin installed successfully."
+
+# Install plugin with Aqua support
+.PHONY: install-plugin-aqua
+install-plugin-aqua:
+	$(MAKE) install-plugin BUILDTAGS=aqua
 
 .PHONY: run
 run:
 	@trivy_version=$$(cat go.mod | grep 'github.com/aquasecurity/trivy v' | awk '{ print $$2}') ;\
 	echo Current trivy version: $$trivy_version ;\
-	go build -ldflags "-s -w -X github.com/aquasecurity/trivy-mcp/pkg/version.TrivyVersion=$${trivy_version}" -o trivy-mcp ./cmd/trivy-mcp/main.go;
-	@./trivy-mcp --version
+	go build $(BUILD_FLAGS) -ldflags "-s -w -X github.com/aquasecurity/trivy-mcp/pkg/version.TrivyVersion=$${trivy_version}" -o trivy-mcp ./cmd/trivy-mcp/main.go;
+	@./trivy-mcp help
 
 .PHONY: lint
 lint:
@@ -67,4 +81,4 @@ bump-manifest:
 	@current_version=$$(cat plugin.yaml | grep 'version' | awk '{ print $$2}' | tr -d '"') ;\
 	echo Current version: $$current_version ;\
 	echo New version: $$NEW_VERSION ;\
-	$(SED) -i -e "s/$$current_version/$$NEW_VERSION/g" plugin.yaml ;\
+	$(SED) -i -e "s/$$current_version/$$NEW_VERSION/g" plugin.yaml ;
