+++ pkg/flag/flag.go
@@ -23,14 +23,6 @@ func AddBaseFlags(cmd *cobra.Command) {
 	cmd.Flags().BoolVarP(&debug, "debug", "d", false, "Enable debug mode")
 }
 
-func AddMcpFlags(cmd *cobra.Command) {
-	cmd.Flags().StringVarP(&transport, "transport", "t", "stdio", "Transport protocol to use (sse or stdio)")
-	cmd.Flags().IntVarP(&port, "port", "p", 8080, "Port to listen on")
-	cmd.Flags().StringVarP(&trivyBinary, "trivy-binary", "", "", "Path to the Trivy binary")
-	// TODO: Uncomment the following line when the Aqua platform is ready
-	//cmd.Flags().BoolVarP(&useAquaPlatform, "use-aqua-platform", "a", false, "Use Aqua platform")
-}
-
 func AddLoginFlags(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&aquaKey, "aqua-key", "", "Aqua key")
 	cmd.Flags().StringVar(&aquaSecret, "aqua-secret", "", "Aqua secret")
